pyproject.toml
[tool.poetry]
name = "blank"
version = "0.1.0"
description = "The description of the package"

authors = ["Eduardo Castro <info@ecdesign.es>"]

readme = "README.md"

repository = "https://github.com/ecDesignStudio/blank"
homepage = "https://github.com/ecDesignStudio/blank"

keywords = ["packaging", "poetry"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.12"
django-cleanup = "^6.0.0"
django-crispy-forms = "^1.14.0"
django-dynamic-preferences = "^1.12.0"
django-modeltranslation = "^0.17.5"
django-render-block = "^0.9.1"
django-rosetta = "^0.9.8"
django-split-settings = "^1.1.0"
python-dateutil = "^2.8.2"
python-decouple = "^3.6"
sorl-thumbnail = "^12.8.0"
mysqlclient = "^2.1.0"
html2text = "^2020.1.16"
cssutils = "^2.4.0"
requests = "^2.27.1"
pytest = "^7.1.1"
django-utils-six = "^2.0"
Pillow = "^9.0.1"
django-ckeditor = "^6.2.0"
django-meta = "^2.0.0"
django-cors-headers = "^3.12.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
django-admin-generator = "^2.2.1"
django-debug-toolbar = "^3.2.4"
django-extensions = "^3.1.5"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
